move bitwise operations
